requirements.txt
cython
fasttext==0.8.3
mecab-python3==0.7
